+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -1,10 +1,9 @@
 <template>
     <div>
         <error-dialog
-            v-if="shouldShowErrorDialog"
             :is-open="shouldShowErrorDialog"
-            :error-code="nonFulfillableError.code"
-            @handle-close="handleErrorDialogClose"
+            :error-code="hasNonFulfillableErrorCode"
+            @close="handleErrorDialogClose"
             @checkout-error-dialog-button-click="handleErrorDialogButtonClick" />
         <div
             v-if="shouldShowSpinner"
@@ -20,6 +19,7 @@
         >
             <alert
                 v-if="genericErrorMessage"
+                ref="errorAlert"
                 type="danger"
                 :class="$style['c-checkout-alert']"
                 :heading="$t('errorMessages.errorHeading')"
@@ -111,6 +111,7 @@ import '@justeat/f-form-field/dist/f-form-field.css';
 
 import { validations } from '@justeat/f-services';
 import { VueGlobalisationMixin } from '@justeat/f-globalisation';
+import VueScrollTo from 'vue-scrollto';
 
 import AddressBlock from './Address.vue';
 import CheckoutHeader from './Header.vue';
@@ -201,6 +202,12 @@ export default {
             type: Number,
             required: false,
             default: 1000
+        },
+
+        spinnerTimeout: {
+            type: Number,
+            required: false,
+            default: 500
         },
 
         authToken: {
@@ -323,6 +330,10 @@ export default {
 
         restaurantMenuPageUrl () {
             return `restaurant-${this.restaurant.seoName}/menu`;
+        },
+
+        hasNonFulfillableErrorCode () {
+            return this.nonFulfillableError && this.nonFulfillableError.code;
         },
 
         eventData () {
@@ -715,6 +726,25 @@ export default {
             this.trackFormInteraction({ action: 'error', error: `error_${error.message}` });
 
             this.genericErrorMessage = messageToDisplay || this.$t('errorMessages.genericServerError');
+
+            this.$nextTick(() => {
+                this.scrollToElement('errorAlert');
+            });
+        },
+
+        /**
+         * Scroll to a ref element in the screen.
+        */
+        scrollToElement (refElement) {
+            const scrollingDurationInMilliseconds = 650;
+
+            const element = this.$refs[refElement]
+                ? this.$refs[refElement].$el
+                : null;
+
+            if (element) {
+                VueScrollTo.scrollTo(element, scrollingDurationInMilliseconds, { offset: -20 });
+            }
         },
 
         /**
@@ -795,7 +825,7 @@ export default {
                 if (this.isLoading) {
                     this.shouldShowSpinner = true;
                 }
-            }, 1000);
+            }, this.spinnerTimeout);
         },
 
         handleErrorDialogClose () {
@@ -803,7 +833,7 @@ export default {
         },
 
         handleErrorDialogButtonClick () {
-            if (this.nonFulfillableError.shouldRedirectToMenu) {
+            if (this.nonFulfillableError && this.nonFulfillableError.shouldRedirectToMenu) {
                 window.location.assign(this.restaurantMenuPageUrl);
             }
 
@@ -887,7 +917,8 @@ export default {
 
 .c-checkout-alert {
     width: $checkout-width;
-    margin: 0 auto;
+    margin-left: auto;
+    margin-right: auto;
 }
 /* If these stay the same then just rename the class to something more generic */
 .c-checkout-submitButton {