+++ packages/components/organisms/f-checkout/src/store/tests/checkout.module.test.js
@@ -7,7 +7,7 @@ import customerAddresses from '../../demo/get-address.json';
 import geoLocationDetails from '../../demo/get-geo-location.json';
 import { mockAuthToken } from '../../components/_tests/helpers/setup';
 import { version as applicationVerion } from '../../../package.json';
-import { VUEX_CHECKOUT_ANALYTICS_MODULE } from '../../constants';
+import { VUEX_CHECKOUT_ANALYTICS_MODULE, DEFAULT_CHECKOUT_ISSUE } from '../../constants';
 
 import {
     UPDATE_AUTH,
@@ -158,9 +158,22 @@ describe('CheckoutModule', () => {
                 expect(state.customer).toEqual(defaultState.customer);
             });
 
-            it('should leave address state empty if no address data is returned from the API.', () => {
+            it('should leave address state empty if no location data is returned from the API.', () => {
                 // Arrange
-                checkoutDelivery.address = null;
+                checkoutDelivery.location = null;
+
+                // Act
+                mutations[UPDATE_STATE](state, checkoutDelivery);
+
+                // Assert
+                expect(state.address).toEqual(defaultState.address);
+            });
+
+            it('should leave address state empty if location data is returned with no address from the API.', () => {
+                // Arrange
+                checkoutDelivery.location = {
+                    address: null
+                };
 
                 // Act
                 mutations[UPDATE_STATE](state, checkoutDelivery);
@@ -450,6 +463,7 @@ describe('CheckoutModule', () => {
                 axios.patch = jest.fn(() => Promise.resolve({
                     status: 200,
                     data: {
+                        isFulfillable: false,
                         issues
                     }
                 }));
@@ -461,6 +475,14 @@ describe('CheckoutModule', () => {
 
                 // Assert
                 expect(axios.patch).toHaveBeenCalledWith(payload.url, payload.data, config);
+            });
+
+            it('should convert an unsupported error into a default error.', async () => {
+                // Act
+                await updateCheckout({ commit, state }, payload);
+
+                // Assert
+                expect(commit).toHaveBeenCalledWith(UPDATE_ERRORS, [{ code: DEFAULT_CHECKOUT_ISSUE, shouldShowInDialog: true }]);
             });
         });