+++ packages/components/organisms/f-checkout/test/specs/component/shared/f-checkout-guest.component.spec.js
@@ -8,7 +8,8 @@ const checkout = new Checkout('organism', 'checkout-component');
 describe('f-checkout component tests', () => {
     beforeEach(() => {
         checkout.withQuery('&knob-Service Type', 'delivery')
-        .withQuery('&knob-Is User Logged In', false);
+        .withQuery('&knob-Is User Logged In', false)
+        .withQuery('&knob-Is Pre Order Warning Displayed, true);
 
         const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
 
@@ -49,7 +50,7 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        checkout.populateCheckoutForm(emailAddress);
+        checkout.populateGuestCheckoutForm(emailAddress);
         checkout.goToPayment();
 
         // Assert
@@ -75,8 +76,8 @@ describe('f-checkout component tests', () => {
         // Assert
         expect(checkout.isOrderTimeDropdownDisplayed()).toBe(true);
         expect(checkout.getOrderTimeOptionText(0)).toBe('As soon as possible');
-        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 00:45');
-        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 01:00');
+        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 01:45');
+        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 02:00');
     });
 
     it('should display a "mobileNumber" error message when an unsupported country code is used in the mobile number field', () => {