+++ packages/components/organisms/f-checkout/test/specs/component/shared/f-checkout.component.spec.js
@@ -6,7 +6,8 @@ const checkout = new Checkout('organism', 'checkout-component');
 describe('f-checkout component tests', () => {
     beforeEach(() => {
         checkout.withQuery('&knob-Service Type', 'delivery')
-                .withQuery('&knob-Is User Logged In', true);
+                .withQuery('&knob-Is User Logged In', true)
+                .withQuery('&knob-Is Pre Order Warning Displayed', true);
 
         const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
 
@@ -79,8 +80,8 @@ describe('f-checkout component tests', () => {
         // Assert
         expect(checkout.isOrderTimeDropdownDisplayed()).toBe(true);
         expect(checkout.getOrderTimeOptionText(0)).toBe('As soon as possible');
-        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 00:45');
-        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 01:00');
+        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 01:45');
+        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 02:00');
     });
 
     it('should prevent a user from writing a note of over 200 characters', () => {