+++ packages/components/organisms/f-header/test/specs/component/desktop/f-header.component.spec.js
@@ -57,9 +57,9 @@ describe('Desktop - f-header component tests', () => {
         header.withQuery('&knob-Locale', formatLocale(expectedLocale));
         header.withQuery('&knob-Show offers link', 'true');
         header.withQuery('&knob-Show delivery enquiry', 'true');
-        const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
-        // Act
+        const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
+            // Act
         header.open(pageUrl);
         header.waitForComponent();
         ['offersLink', 'userAccount', 'help', 'countrySelector'].forEach(link => {
@@ -81,6 +81,26 @@ describe('Desktop - f-header component tests', () => {
         header.open(pageUrl);
         header.waitForComponent();
         ['userAccount', 'help', 'countrySelector'].forEach(link => {
+            // Assert
+            expect(header.isNavigationLinkDisplayed(link)).toBe(true);
+            expect(header.isNavigationLinkDisplayed('delivery')).toBe(false);
+        });
+    });
+
+    forEach(['it', 'es', 'dk', 'no'])
+    .it('should display the below navigation links', expectedLocale => {
+        // Arrange
+        const headerData = {
+            locale: expectedLocale,
+            offers: true,
+            delivery: true
+        };
+
+        ['userAccount', 'help', 'countrySelector'].forEach(link => {
+            // Act
+            header.open(headerData);
+            header.waitForComponent();
+
             // Assert
             expect(header.isNavigationLinkDisplayed(link)).toBe(true);
             expect(header.isNavigationLinkDisplayed('offersLink')).toBe(false);
@@ -113,6 +133,7 @@ describe('Desktop - f-header component tests', () => {
         header.withQuery('&knob-Show offers link', 'true');
         header.withQuery('&knob-Show delivery enquiry', 'true');
         const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
+
 
         // Act
         header.open(pageUrl);