+++ packages/components/organisms/f-registration/test/specs/component/shared/f-registration.component.spec.js
@@ -81,8 +81,8 @@ const registration = new Registration('organism', 'registration-component');
         const userInfo = {
             firstName: 'abcdefghijabcdefghijabcdefghijabcdefghijabcdefghijabcdefghij',
             lastName: 'abcdefghijabcdefghijabcdefghijabcdefghijabcdefghijabcdefghij',
-            email: 'ashton.adamms+jetest@just-eat.com',
-            password: 'llanfairpwllgwyngyllgogerychwyr'
+            email: 'joe@test.com',
+            password: 'abcdefghijabcdefghijabcdefghi'
         };
 
         // Act