+++ packages/components/organisms/f-status-banner/test-utils/component-objects/f-statusBanner.selectors.js
@@ -1,0 +1,7 @@
+export const COMPONENT = '[data-test-id="main-banner-container"]';
+export const ADDRESS_BOX = '[data-test-id="address-box-input"]';
+export const BUTTON = '[data-test-id="find-restaurants-button"]';
+export const ERROR_MESSAGE = '[id="errorMessage"]';
+export const BANNER_IMAGE = '[data-test-id="banner-image"]';
+export const BANNER_HEADING = '[data-test-id="banner-search-box"] h1';
+export const BANNER_SUBHEADING = '[data-test-id="banner-search-box"] h2';